Updated Hillary
Set Nickname To "Hillary"
Set Description To "Hillary is a sweet doe out of a special doe (1st Lady) who was the 1st after about 20 bucklings the year she was born. I had taken too many doeling reservations that year and had to sell Lady. She did not work out in the home she was originally placed in and my friend Annette took her for me until I could make arrangements to bring her home. In the meantime, we bred her at Annette's and the result was a stellar doe Seagal. Both Lady and Seagal went down to sunny CA to contribute to the budding Blackberry herd. I kept this doeling and she has done us proud. Hillary is a huge milker who is a dream on the milk stand and I count on her amazing milk late into the season for my morning coffee! "
Set Normalize Id To "AMN06647"