Updated Soaring Heart's Singin in the Rain
Set Description To "Singin in the Rain is out of my all time favorite doe Purple Rain and by Mal Maison a buck I brought back as a 7 year old from WY. While originally I wanted a buck replacement (and Purple did comply with one), my herd situation downsizing didn't make sense to keep any young bucks at this time. Luckily we found a local farm where he will be easy to either use, or bring back in a few years once he has served his current farm. Instead we kept this doeling. Purple once again gifted us with quads! Since I won't be breeding any of this years doelings, we have a year or so to evaluate a crop of them. THIS one will stay for sure! This is not a great picture of Rain's beautiful breed character, but the best view of overall."