Updated Petite
Set Nickname To "Petite"
Set Description To "Petite is a lovely correct doe out of Soaring heart's Sienna Princess *P(now at Mini Love Farm where she earned her milk star that is not reflected on Petite's registration paperwork) and by  Soaring Heart's Mal Maison *B. She is dual registered with both TMGR and MDGA and will be on milk test in 2023 as a first freshener. The picture is from Petite as baby and will be updated soon with udder pics. "
Set Normalize Id To "AMN14333"
Set Date Of Birth To "07/26/2021"