Updated Pearlyna
Set Nickname To "Pearlyna"
Set Description To "Pearlyna is named for a resort in Ukraine, born in the month that Russia invaded and begin an horrific attack on this country, it was one of those timely names. Like her Brother King, Pearlyna is precious as the offspring of two deceased parents we had great faith in. We look forward to seeing this doeling freshen as yearling with Versailles."
Set Normalize Id To "DN000302014149"
Set Date Of Birth To "1/12/2022"