Updated Sexy
Set Nickname To "Sexy"
Set Date Of Birth To "3/20/2022"
Set Normalize Id To "BN000302014191"
Set Description To "Sexy is a beautiful 6th generation buck. He is our Marvin replacement and of all the young bucks, he was all about mastering his new job! I had the pleasure (truly) of showing him in the May in the Motherlode show last spring. He was a complete little gentleman and convinced the judge to place him 1st in the Jr buck division, unfortunately the buck class was combined Jr/Sr and the Sr bucks won out for GCH and RGCH. We look forward to seeing his kids with Rosemary, Beryl Petite and Calamity this year. "