Updated Nurse King
Set Nickname To "Nurse King"
Set Description To "Nurse King is named for a dear friend and nursing colleage. He is precious as a survivor of his very special dam Footloose Acres Pearl and Dr David both deceased. He has great promise and pairs some of our strongest lines. King is polled and the first chocolate for us. We look forward to his kids with Tundra (sold but with first choice doeling), And Fille."
Set Date Of Birth To "3/12/2022"
Set Normalize Id To "BN000302014150"