Updated Versailles
Set Nickname To "Versailles"
Set Description To "Versaills is the son of our 1st lady Hillary and Mal Maison. Champions and proven milk production both sire and dam. He was a later kid and is just begining to figure things out. He has a very cute pompador and is a quiet low key dude like his dad. We look forward to kids out of Pepper and Pearlyna this year. "
Set Normalize Id To "BN000302014194"
Set Date Of Birth To "3/30/2022"